feat(vue): Vuex and router + move dashboard to controller specific
