+++ package.json
@@ -19,10 +19,13 @@
         "build": "encore production --progress"
     },
     "dependencies": {
+        "axios": "^0.24.0",
         "eslint": "^8.6.0",
         "tmi.js": "^1.8.5",
         "vue": "^2.6.14",
         "vue-loader": "^15.9.8",
-        "vue-template-compiler": "^2.6.14"
+        "vue-router": "^3.5.3",
+        "vue-template-compiler": "^2.6.14",
+        "vuex": "^3.6.2"
     }
 }
